Refactor edit files
