+++ frontend/src/functionality/CreateSchedule/utils/Hooks/useCreateTimeBlock.js
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { api } from '../../api.js';
+import { api } from '../../../../api.test.js';
 
 /**
  * Creates one or more time block via the API.
